Took out logic from table component and placed it to pages leaving table component non-duplicated
