+++ src/components/Tables/MainTable/TableRows/Rows/WarehouseRow.jsx
@@ -1,9 +1,8 @@
-import React, { useEffect, useState } from 'react';
+import React, { memo, useEffect, useState } from 'react';
 import TableRow from '@mui/material/TableRow';
 import { MDBBtn, MDBIcon } from 'mdb-react-ui-kit';
 
 import TableCell from '@mui/material/TableCell';
-
 
 import { PRODUCT_WAREHOUSE_QUANTITY } from '../../../consts';
 
@@ -14,7 +13,7 @@ import '../../../TableStyles/TableRows/OneRow/OneRow.scss';
 
 const relations = new Relations(PRODUCT_WAREHOUSE_QUANTITY);
 
-const OneRow = ({ row, deleteFunc, editItem }) => {
+const WarehouseRow = memo(({ row, deleteFunc, editItem }) => {
   const [showButtons, setShowButtons] = useState(false);
   const [quantity, setQuantity] = useState(0);
 
@@ -62,6 +61,6 @@ const OneRow = ({ row, deleteFunc, editItem }) => {
 
     </TableRow>
   );
-};
+});
 
-export default OneRow;
+export default WarehouseRow;
